Initial bundle size reduction and bugfixes
