+++ web/src/components/Root.jsx
@@ -17,7 +17,7 @@ const Root = styled('div')({
 
 const FooterContainer = styled('div')(({ theme }) => ({
   paddingTop: theme.spacing(10),
-  position: 'sticky',
+  position: 'static',
   left: 0,
   overflow: 'auto',
 }));
