+++ web/src/components/TopBar/index.ts
@@ -1,0 +1,3 @@
+import TopBar from './TopBar';
+
+export default TopBar;
