+++ web/src/pages/_document.tsx
@@ -25,7 +25,7 @@ export default function MyDocument({ emotionStyleTags }: MyDocumentProps) {
     <Html lang="en" className={roboto.className}>
       <Head>
         {/* PWA primary color */}
-        <meta name="theme-color" content={(theme as any).colorSchemes.dark.palette.main} />
+        <meta name="theme-color" content={theme.colorSchemes.dark.palette.primary.main} />
         <link rel="shortcut icon" href="/favicon.ico" />
         <meta name="emotion-insertion-point" content="" />
         {emotionStyleTags}
